+++ packages/react/src/components/form/authenticating/password.tsx
@@ -34,7 +34,10 @@ const PasswordRecoveryPrompt = ({
   const { text } = useConfiguration();
 
   return (
-    <div className={styles.passwordRecoveryPrompt}>
+    <div
+      data-testid="sid-form-authenticating-recover-prompt"
+      className={styles.passwordRecoveryPrompt}
+    >
       <Text
         variant={{ size: "sm", color: "tertiary", weight: "semibold" }}
         t="authenticating.verifyPassword.recover.prompt"
@@ -42,7 +45,7 @@ const PasswordRecoveryPrompt = ({
       <LinkButton
         className={sprinkles({ marginLeft: "1" })}
         type="button"
-        testId="sid-form-authenticating-retry-button"
+        testId="sid-form-authenticating-recover-button"
         onClick={onRecoverClick}
       >
         {text["authenticating.verifyPassword.recover.cta"]}